+++ app/assets/javascripts/konacha/parent.js
@@ -1,10 +1,15 @@
 window.onload = function () {
+  mocha.suite.suites.sort(function (a, b) {
+    return a.path.localeCompare(b.path);
+  });
+
   var iframes = document.getElementsByTagName('iframe');
   for (var i = 0; i < iframes.length; ++i) {
     if (!iframes[i].contentWindow.mocha) {
       (function (path) {
         mocha.suite.addTest(new Mocha.Test(path, function () {
-          throw new Error("Failed to load " + path + ". Perhaps it failed to compile?");
+          throw new Error("Failed to load " + path + ".\n" +
+                          "Perhaps it failed to compile? Check the rake output for errors.");
         }));
       })(iframes[i].getAttribute("data-path"));
     }