+++ config/routes.rb
@@ -1,6 +1,5 @@
 Konacha::Engine.routes.draw do
-  match '/iframe'       => 'specs#iframe', :as => 'iframe'
-  match '/iframe/*path' => 'specs#iframe'
-  match '/'             => 'specs#parent', :as => 'parent'
+  match '/iframe/*name' => 'specs#iframe'
+  match '/'             => 'specs#parent'
   match '*path'         => 'specs#parent'
 end