+++ spec/views/specs/iframe.html.erb_spec.rb
@@ -15,23 +15,21 @@ describe "konacha/specs/iframe" do
 
   def spec_double(asset_name, dependencies = [])
     asset_double(asset_name, dependencies)
-    double("spec called '#{asset_name}'", :asset_name => asset_name)
+    double("spec called '#{asset_name}'", :asset_name => asset_name, :path => "#{asset_name}.js")
   end
 
   let(:dependency) { asset_double("dependency") }
 
-  it "renders a script tag for each spec in @specs" do
-    assign(:specs, [spec_double("a_spec"),
-                    spec_double("b_spec")])
+  it "renders a script tag for @spec" do
+    assign(:spec, spec_double("a_spec"))
 
     render
 
     rendered.should have_selector("script[src='/assets/a_spec.js?body=1']")
-    rendered.should have_selector("script[src='/assets/b_spec.js?body=1']")
   end
 
   it "renders a script tag for a spec's dependencies" do
-    assign(:specs, [spec_double("spec", [dependency])])
+    assign(:spec, spec_double("spec", [dependency]))
 
     render
 
@@ -39,20 +37,11 @@ describe "konacha/specs/iframe" do
     rendered.should have_selector("script[src='/assets/spec.js?body=1']")
   end
 
-  it "renders only one script tag for common dependencies" do
-    assign(:specs, [spec_double("a_spec", [dependency]),
-                    spec_double("b_spec", [dependency])])
-
-    render
-
-    rendered.should have_selector("script[src='/assets/dependency.js?body=1']", :count => 1)
-  end
-
   it "renders only one script tag for dependencies of dependencies" do
     dependency_a = asset_double("dependency_a")
     dependency_b = asset_double("dependency_b", [dependency_a])
 
-    assign(:specs, [spec_double("a_spec", [dependency_a, dependency_b])])
+    assign(:spec, spec_double("a_spec", [dependency_a, dependency_b]))
 
     render
 
@@ -61,11 +50,21 @@ describe "konacha/specs/iframe" do
   end
 
   it "render the stylesheets" do
+    assign(:spec, spec_double("a_spec"))
     assign(:stylesheets, %w(foo bar))
+    assign(:specs, [])
 
     render
 
     rendered.should have_selector("link[href='/assets/foo.css']")
     rendered.should have_selector("link[href='/assets/bar.css']")
+  end
+
+  it "includes a path data attribute" do
+    assign(:spec, spec_double("a_spec"))
+
+    render
+
+    rendered.should have_selector("[data-path='a_spec.js']")
   end
 end